Fix timed spells' timers not displaying correctly

+++ src/main/java/com/minelittlepony/unicopia/ability/magic/spell/TimedSpell.java
@@ -29,7 +29,7 @@ public interface TimedSpell extends Spell {
         }
 
         public float getPercentTimeRemaining(float tickDelta) {
-            return MathHelper.lerp(tickDelta, prevDuration, duration) / maxDuration;
+            return MathHelper.lerp(tickDelta, prevDuration, duration) / (float)maxDuration;
         }
 
         public int getTicksRemaining() {
